requirements.txt
# PDF-to-Chatbot Platform Requirements

# Core Framework
streamlit>=1.28.0

# AI/ML
openai>=1.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.0

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Database
supabase>=2.0.0
postgrest>=0.13.0

# Utilities
python-dotenv>=1.0.0
python-multipart==0.0.6